requirements.txt
fastapi
uvicorn[standard]
pydantic
google-cloud-storage
google-cloud-firestore
google-cloud-tasks
google-auth
python-multipart
